汎用じゃない関数を移動
