+++ app/routers/guild/vc_signal/vc_signal.py
@@ -3,14 +3,14 @@ from fastapi.responses import RedirectResponse,JSONResponse
 from starlette.requests import Request
 from fastapi.templating import Jinja2Templates
 
-from typing import List,Dict,Any
+from typing import List,Dict,Any,Tuple
+from itertools import groupby,chain
 
 from base.aio_req import (
     aio_get_request,
     return_permission,
     discord_oauth_check,
     get_profile,
-    sort_discord_vc_channel,
     sort_channels
 )
 from model_types.discord_type.guild_permission import Permission
@@ -582,3 +582,95 @@ async def chenge_permission_check(
     else:
         # 変更不可
         return False
+
+async def sort_discord_vc_channel(
+    all_channel:List
+) -> Tuple[List,List,List]:
+    # 親カテゴリー格納用
+    position = []
+    # ソート後のチャンネル一覧
+    all_channel_sort = []
+
+    # レスポンスのJSONからpositionでソートされたリストを作成
+    sorted_channels = sorted(all_channel, key=lambda c: c['position'])
+
+    # parent_idごとにチャンネルをまとめた辞書を作成
+    channel_dict = {}
+
+    for parent_id, group in groupby(
+        sorted_channels,
+        key=lambda c: c['parent_id']
+    ):
+        if parent_id is None:
+            # 親カテゴリーのないチャンネルは、キーがNoneの辞書に追加される
+            parent_id = 'None'
+
+        # キーがまだない場合、作成(同時に値も代入)
+        if channel_dict.get(str(parent_id)) == None:
+            channel_dict[str(parent_id)] = list(group)
+        # キーがある場合、リストを取り出して結合し代入
+        else:
+            listtmp:List = channel_dict[str(parent_id)]
+            listtmp.extend(list(group))
+
+            # リスト内包記法でボイスチャンネルとカテゴリー以外は除外
+            listtmp = [
+                tmp
+                for tmp in listtmp
+            ]
+            channel_dict[str(parent_id)] = listtmp
+            # リストを空にする
+            listtmp = list()
+
+    # 親カテゴリーがある場合、Noneから取り出す
+    for chan in channel_dict['None'][:]:
+        if chan['type'] == 4:
+            position.append(chan)
+            channel_dict['None'].remove(chan)
+
+    # 辞書を表示
+    position_index = 0
+
+    # 親カテゴリーの名前をリスト化
+    extracted_list = [d["name"] for d in position]
+    # カテゴリーに属しないチャンネルが存在する場合
+    if len(channel_dict['None']) != 0:
+        # 配列の長さをカテゴリー数+1にする
+        all_channels = [{}] * (len(extracted_list) + 1)
+        vc_channels = [{}] * (len(extracted_list) + 1)
+    else:
+        all_channels = [{}] * len(extracted_list)
+        vc_channels = [{}] * len(extracted_list)
+
+    for parent_id, channel in channel_dict.items():
+        # カテゴリー内にチャンネルがある場合
+        if len(channel) != 0:
+            for d in position:
+                # カテゴリーを探索、あった場合positionを代入
+                if d['id'] == channel[0]['parent_id']:
+                    position_index = d['position']
+                    break
+        else:
+            position_index = len(extracted_list)
+
+        if len(channel) != 0:
+            # 指定したリストの中身が空でない場合、空のリストを探す
+            while len(all_channels[position_index]) != 0:
+                if len(extracted_list) == position_index:
+                    position_index -= 1
+                else:
+                    position_index += 1
+
+            # 指定した位置にカテゴリー内のチャンネルを代入
+            all_channels[position_index] = channel
+            vc_channels[position_index] = channel
+
+            # 先頭がカテゴリーでない場合
+            if channel[0]['parent_id'] != None:
+                # 先頭にカテゴリーチャンネルを代入
+                all_channels[position_index].insert(0,d)
+
+    # list(list),[[],[]]を一つのリストにする
+    all_channel_sort = list(chain.from_iterable(all_channels))
+
+    return all_channel_sort,all_channels,vc_channels
